+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Procedure when bumping the version number:
 
 ## Unreleased
 
+### Added
+- `Frame::set_title`
+
 ## v0.2.0 - 2023-08-31
 
 ### Changed